Implemented OpenAL sound system and handler.
Added a new sound system that plays through OpenAL and can handle either a
static mode in which an entire sound file (only WAV is currently supported) can
be loaded into memory and played, or a streaming mode, in which the audio is
buffered from the file via. an audio thread which reduces the memory footprint.

OpenAL automatically adds in support for 3D audio too.