Implemented gamepad input system.
Input can now be received through the use of up to eight concurrent gamepads.
They should be able to be hot-plugged during runtime but it hasn't been tested
thoroughly yet. There is still a TODO for the SDL remapping event too which is
currently just a stub switch case. Overall works well.