Implemented mutex and thread for Win32 and fixed project for win compile.
Implemented the Win32 variants of Smutex and Sthread. Some things had to
change such as the renaming of some vars in Scamera to avoid compile
conflicts but more importantly, the nanosleep function was removed for
Sthread because there isn't any good solution for nanosecond resolution
sleep on Windows, and given the inaccuracy of sleeping for such little
time (on both platforms) it makes no sense to keep it.