Added Win32 code for TCP sockets.
Win32 code was added for TCP sockets, and given Winsock2 is based on Unix
sockets, not much had to be changed, so tcp.c is a cross-platform file (for
now).

socket_posix.c and socket_win32.c were added to deal with initialising the
Winsock library and so for now the POSIX file contains only empty functions.