Added display mode functions and added camera hook for window.
Two new functions were added to get the size of any display on the
operating system in window.h.

The camera can now hook itself exclusively onto a window whereby it will
be automatically adjusted to match any change in width/height/aspect
ratio of the window should it resize.