Moved dr_libs and stb folders into include/ on install.
When `make install' was called, the dr_libs and stb folders were not
moved into the include/ folder thus rendering the library uncompilable.
This commit fixes the issue and additionally fixes some issues
concerning the incorrect copying of header files on `make install' and
not correctly removing all files on `make uninstall'.