Made drwav opaque and reverted commit ccfcb5e4.
drwav is now implemented in src/audio/sound.c rather than in the header
so that programs that use sticky no longer rely on the dr_libs (or stb)
headers being visible in order to compile. As a result, the commit that
changes the directory layout for these headers has been reverted in
favour of this better solution.

The drwav struct is allocated on the heap now too.