Replaced Spencil by drawing functions that interact directly with the window.
Spencil was somewhat of a patchjob - all of the draw functions have been
integrated directly with Swindow in a much nicer solution which no
longer necessitates the usage of 'yet another' Spencil struct and managing
the memory, &c.

The same will be done shortly with Sfont's drawing functions, using the struct
only for storage of the actual font data instead, and in the future, so shall
the draw code for Smodel, &c, leading to a unified rendering file.