Added font extents function, changed some float->Sfloat and fixed math min/max.
There is a new function S_font_get_extents which will return the amount
of space required on screen to draw text in a given font at a given
size. This is useful for writing graphical libraries that want to make
use of font metrics, &c.

The S_min and S_max functions were not working correctly for floats so
they have been changed (fixme?)