Added a custom Quicksort algorithm and replaced the qsort ref in Sfont with it.
A new 'algorithm' module was created with a new set of functions for
running a sequential Quicksort on an array. There are currently two
versions available: a regular sorting function that takes a comparison
function, and an inline version in the form of a macro that will run a
comparison in-place.