Added a version update script.
Added a new tool script that will update the library version. Windows
builds do not respect the VERSION var inside of vars.mk, and instead use
their own variables in sticky.vcxproj. This tool removes the headache of
remembering where each copy of the same variable is defined, so instead
we can use this tool to update it globally for us.