Fixed invalid memory access due to incorrect array size.
The address sanitiser has found a problem where the draw meshes are
generated with incorrect bounds on a vertex array. This has been fixed
by using sizeof to find the size at compile-time rather than hardcoding
the value and ultimately forgetting to change it whenever a change
occurs to the array.