Properly fixed memtrace thread collisions.
Added THREAD_LOCAL to some vars yesterday but it wasn't a proper fix.
I've modified the way that mutexes are created and destroyed: the
underlying functions _S_mutex_new and _S_mutex_delete take a new
'reporting' parameter which will either add to the memory frame when
created or be ignored. The only time they do not report is within
memtrace, because theres a circular dependency on the mutex existing
before adding frames, and the frame being created for the mutex to
exist. Confusing? Yes.

To put it simply, instead of thread-local storage for memtrace, we use a
mutex to force each thread to wait for modification allowance.