Added framebuffer capabilities.
Added a framebuffer to the renderer which allows for scenes to be
rendered to locations other than the window. The colour layer is stored
as a texture and depth/stencil layers are stored as renderbuffers for
speed. The code is still in development and a few things need to be fixed,
namely:

- Setting the draw colour with a different framebuffer attached should
  be reset once the framebuffer is detached.
- The draw code in draw.c is window-orientated and needs to be
  reconciliated for framebuffer use.