Disabled OpenMP on macOS.
GitHub actions are now complaining for macOS builds because brew changed
the way that OpenMP is used. By default, the macOS build of clang does
not support OpenMP and the library does not ship with the system, so
OpenMP will instead be disabled to confer with Apple but remain turned
on for Windows and Linux systems.